fix routes, update and delete posts
