+++ controllers/api/postRoutes.js
@@ -2,65 +2,6 @@ const router = require("express").Router();
 const { Post } = require("../../models");
 const withAuth = require("../../utils/auth");
 
-// display all posts on dashboard
-router.get("/", withAuth, async (req, res) => {
-  try {
-    // Get all projects and JOIN with user data
-    const postData = await Post.findAll({
-      where: {
-        user_id: req.session.user_id,
-      },
-    });
-
-    // Serialize data so the template can read it
-    const posts = postData.map((post) => post.get({ plain: true }));
-
-    // Pass serialized data and session flag into template
-    res.render("dashboard", {
-      posts,
-      logged_in: req.session.logged_in,
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-// route for creating a new blog post
-router.get("/postCreate", withAuth, async (req, res) => {
-  res.render("postCreate", {
-    logged_in: req.session.logged_in,
-  });
-});
-
-router.post("/postCreate", withAuth, async (req, res) => {
-  try {
-    const newPost = await Post.create({
-      ...req.body,
-      user_id: req.session.user_id,
-    });
-
-    res.status(200).json(newPost);
-  } catch (err) {
-    res.status(400).json(err);
-  }
-});
-
-// route for viewing a sinlge blog post
-router.get("/postUpdate/:id", withAuth, async (req, res) => {
-  try {
-    const postData = await Post.findByPk(req.params.id, {});
-
-    const post = postData.get({ plain: true });
-
-    res.render("postUpdate", {
-      ...post,
-      logged_in: req.session.logged_in,
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
 // route for creating a new blog post
 router.post("/", withAuth, async (req, res) => {
   try {
@@ -98,7 +39,7 @@ router.delete("/:id", withAuth, async (req, res) => {
 
 // route for updating an existing blog post
 // router.put("/:id", withAuth, async (req, res) => {
-router.put("/postUpdate/:id", async (req, res) => {
+router.put("/:id", async (req, res) => {
   try {
     const postData = await Post.update(req.body, {
       where: {
