+++ controllers/homeRoutes.js
@@ -21,7 +21,7 @@ router.get("/", async (req, res) => {
     // res.render('homepage', {posts});
     res.render("homepage", {
       posts,
-      // logged_in: req.session.logged_in
+      logged_in: req.session.logged_in
     });
   } catch (err) {
     res.status(500).json(err);
