+++ controllers/index.js
@@ -1,8 +1,10 @@
 const router = require('express').Router();
 const apiRoutes = require('./api');
 const homeRoutes = require('./homeRoutes');
+const dashboardRoutes =  require('./dashboardRoutes')
 
 router.use('/api', apiRoutes);
+router.use('/dashboard', dashboardRoutes);
 router.use('/', homeRoutes);
 
 module.exports = router;
