+++ public/js/login.js
@@ -15,7 +15,7 @@ const loginFormHandler = async (event) => {
 
     if (response.ok) {
       // If successful, redirect the browser to the dashboard page
-      document.location.replace("/api/posts");
+      document.location.replace("/dashboard");
     } else {
       alert(response.statusText);
     }
