+++ public/js/postCreate.js
@@ -12,7 +12,6 @@ const postFormHandler = async (event) => {
     });
 
     if (response.ok) {
-      // If successful, redirect the browser to the dashboard page
       document.location.replace("/api/posts");
     } else {
       alert(response.statusText);
