+++ public/js/signup.js
@@ -13,7 +13,7 @@ const signupFormHandler = async (event) => {
       });
   
       if (response.ok) {
-        document.location.replace("/api/posts");
+        document.location.replace("/dashboard");
       } else {
         alert(response.statusText);
       }
