+++ views/dashboard.handlebars
@@ -3,10 +3,10 @@
   <div class="row mb-4 project">
     <div class="col-md-5">
       <h2>
-        <a href="/api/posts/postUpdate/{{post.id}}">{{post.title}}</a>
+        <a href="/dashboard/update/{{post.id}}">{{post.title}}</a>
       </h2>
     </div>
   </div>
 {{/each}}
 
-<a href="/api/posts/postCreate"><button>Create new post</button></a>
+<a href="/dashboard/postCreate"><button>Create new post</button></a>
