+++ views/layouts/main.handlebars
@@ -5,7 +5,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Tech Blog</title>
-  <link rel="stylesheet" href="/css/jass.css">
+ <!-- CSS only -->
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
   <link rel="stylesheet" href="/css/style.css">
 </head>
 
@@ -16,15 +17,13 @@
         <a href="/">Tech Blog!!</a>
       </h1>
       <nav>
+        <a href="/">homepage</a> |
         {{#if logged_in}}
         <!-- Conditionally render login or logout links -->
-        <a href="/">homepage</a> |
-        <a href="/api/posts">dashboard</a> |
+        <a href="/dashboard">dashboard</a> |
         <button class="no-button" id="logout">logout</button>
         {{else}}
-        <a href="/">homepage</a> |
-        <a href="/api/posts">dashboard</a> |
-        <a href="/login">login</a>
+         <a href="/login"><button type="button">login/signup</button></a>
         {{/if}}
       </nav>
     </header>
