+++ views/login.handlebars
@@ -17,7 +17,7 @@
     </form>
   </div>
   <div>
-    <a href="/signup">Signup instead</a>
+    <p>no account? <a href="/signup">Signup instead</a></p>
   </div>
 </div>
