+++ views/postUpdate.handlebars
@@ -11,10 +11,12 @@
     <div class="form-group">
       <label for="update-content">content:</label>
       <textarea class="form-input" type="text" id="update-content">{{content}}</textarea>
-      {{!-- <input class="form-input" type="hidden" id="post_id" value="{{id}}" /> --}}
+      <input class="form-input" type="hidden" id="post_id" value="{{id}}" />
     </div>
     <div class="form-group">
       <button class="btn btn-primary" type="submit">update my post</button>
-      <button class="btn btn-primary" type="submit">delete my post</button>
+      <button class="btn btn-primary" type="button" id="deleteBtn">delete my post</button>
     </div>
   </form>
+
+  <script src="/js/postUpdate.js"></script>
