+++ views/signup.handlebars
@@ -1,19 +1,22 @@
-  <div class="col-md-6">
-    <h2>Signup</h2>
+<div class="col-md-6">
+  <h2>Signup</h2>
 
-    <form class="form signup-form">
-      <div class="form-group">
-        <label for="username-signup">username:</label>
-        <input class="form-input" type="text" id="username-signup" />
-      </div>
-      <div class="form-group">
-        <label for="password-signup">password:</label>
-        <input class="form-input" type="password" id="password-signup" />
-      </div>
-      <div class="form-group">
-        <button class="btn btn-primary" type="submit">signup</button>
-      </div>
-    </form>
+  <form class="form signup-form">
+    <div class="form-group">
+      <label for="username-signup">username:</label>
+      <input class="form-input" type="text" id="username-signup" />
+    </div>
+    <div class="form-group">
+      <label for="password-signup">password:</label>
+      <input class="form-input" type="password" id="password-signup" />
+    </div>
+    <div class="form-group">
+      <button class="btn btn-primary" type="submit">signup</button>
+    </div>
+  </form>
+  <div>
+    <p>already signed up? <a href="/login">login instead</a></p>
   </div>
+</div>
 
 <script src="/js/signup.js"></script>
